Longest Uncommon Subsequence II
Given an array of strings strs, return the length of the longest uncommon subsequence between them. If the longest uncommon subsequence does not exist, return -1.

An uncommon subsequence between an array of strings is a string that is a subsequence of one string but not the others.

A subsequence of a string s is a string that can be obtained after deleting any number of characters from s.

    For example, "abc" is a subsequence of "aebdc" because you can delete the underlined characters in "aebdc" to get "abc". Other subsequences of "aebdc" include "aebdc", "aeb", and "" (empty string).

 

Example 1:

Input: strs = ["aba","cdc","eae"]
Output: 3

Example 2:

Input: strs = ["aaa","aaa","aa"]
Output: -1

 

Constraints:

    1 <= strs.length <= 50
    1 <= strs[i].length <= 10
    strs[i] consists of lowercase English letters.

Here we have very small input dimensions, so we can directly find all substrings for all words in O(2^k * k * n) and find the longest one which is unique.

However, there is smarter solution: note, that the longest uncommon subsequence will always be one of our words: because if it is not, if we add some letter to it, it also will be uncommon subsequence. So, what we need to do now is to sort words given their lengths in decreasing order. Then for every word in words iterate over all other words and check if given word is subsequence of some other word. We can check if one word is subsequence of another in linear time, using 0392 Is Subsequence with two pointers.

Complexity

Overall time complexity is O(n^2k) to check each of word and compare it with each other word. Space complexity is O(n*k) to keep all sorted words.