Patching Array
Given a sorted integer array nums and an integer n, add/patch elements to the array such that any number in the range [1, n] inclusive can be formed by the sum of some elements in the array.

Return the minimum number of patches required.

 

Example 1:

Input: nums = [1,3], n = 6
Output: 1
Explanation:
Combinations of nums are [1], [3], [1,3], which form possible sums of: 1, 3, 4.
Now if we add/patch 2 to nums, the combinations are: [1], [2], [3], [1,3], [2,3], [1,2,3].
Possible sums are 1, 2, 3, 4, 5, 6, which now covers the range [1, 6].
So we only need 1 patch.

Example 2:

Input: nums = [1,5,10], n = 20
Output: 2
Explanation: The two patches can be [2, 4].

Example 3:

Input: nums = [1,2,2], n = 5
Output: 0

 

Constraints:

    1 <= nums.length <= 1000
    1 <= nums[i] <= 104
    nums is sorted in ascending order.
    1 <= n <= 231 - 1

Solution 1

Let as keep all possible numbers we can get as list of intervals, for example 0, 1, 2, 4, 5, 12 is [[0, 2], [4, 5], [12, 12]]. Then when we add new number we can merge our intervals, using idea of Problem 0056. What numbers we need to add next? We need to add number 3 in our example, the smallest possible number, which are not in our array.

Complexity

Complexity is O(m * log n), where m <= 10000 is the biggest value of num. The idea is that in the beginning we can have not more than m//2 intervals and then on each iteration number of intervals can not increase or increase not too fast (I do not have strict proof at the moment). Also number of iterations can not too big, because it is enough log n patches always.

Solution 2

We can improve this solution, if we upgrade the idea of smallest number which we do not have yet.

Example: nums = [1, 2, 4, 13, 43]. We go form the smallest number add them one by one and always keep the possible range of numbers we can get in the form [0, k], that is, it is not union of intervals, but only one interval. Let us go through this example.

    We did not choose any numbers yet, so we have interval [0, 0].
    Are we missing something if we take next number 1? No, let us take it, now we have interval [0, 1].
    Are we missing something if we take next number 2? No, let us take it, now we have interval [0, 3].
    Are we missing something if we take next number 4? No, let us take it, now we have interval [0, 7].
    Are we missing something if we take next number 13? Yes, we missing! What we need to take next patch equal to 8. So now, our interval is [0, 15].
    Our next missing number 16 is more, than next number in array: 13. So if fact, we can create [0, 28].
    Our next missing number is 29, it is less than 43, so we add 29 as patch. and we have interval [0, 57].

Complexity

Complexity of this solution is O(n).