Orderly Queue
You are given a string s and an integer k. You can choose one of the first k letters of s and append it at the end of the string..

Return the lexicographically smallest string you could have after applying the mentioned step any number of moves.

 

Example 1:

Input: s = "cba", k = 1
Output: "acb"
Explanation: 
In the first move, we move the 1st character 'c' to the end, obtaining the string "bac".
In the second move, we move the 1st character 'b' to the end, obtaining the final result "acb".

Example 2:

Input: s = "baaca", k = 3
Output: "aaabc"
Explanation: 
In the first move, we move the 1st character 'b' to the end, obtaining the string "aacab".
In the second move, we move the 3rd character 'c' to the end, obtaining the final result "aaabc".

 

Constraints:

    1 <= k <= s.length <= 1000
    s consist of lowercase English letters.

/*

Before Jump to the problem let's discuss some cases which may tell us how to proceed to this problem.

EXAMPLE->1: s="bcabc" k=1

now pick index 0 character 'b' and append at the end of the string. s="cabcb"

again pick index 0 character 'c' and append it then s="abcbc".

This modified string is the lexicographically smallest among all the permutations possible by k-moves any number of times.

So,when k is given k=1 , Then we observe that

we are allowed to rotate the string any number of time by 1 character [Left Rotation]

So just apply brute force method to check which rotated string gives us the desired output.

EXAMPLE->2: s="bcaaaaa" k=2

When we are given k==2 in other word it gives us the freedom to swap any two adjacent characters . And obvoiusly for our sake we sort two characters lexicographically at each move . As a result, second smallest character comes after smallest character and the third one comes after second smallest character.

So required transformation looks like---->

"bcaaaaa" --- "caaaaab" --- "aaaaabc".

So as a whole Just we can sort the string and return it.

In a same manner for k>2 alse we can do the above mentioned steps (because the steps we follow for k==2 , we can repeat the same for k>2 also).

*/