Reverse Linked List
Given the head of a singly linked list, reverse the list, and return the reversed list.

 

Example 1:
https://assets.leetcode.com/uploads/2021/02/19/rev1ex1.jpg
Input: head = [1,2,3,4,5]
Output: [5,4,3,2,1]

Example 2:
https://assets.leetcode.com/uploads/2021/02/19/rev1ex2.jpg
Input: head = [1,2]
Output: [2,1]

Example 3:

Input: head = []
Output: []

 

Constraints:

    The number of nodes in the list is the range [0, 5000].
    -5000 <= Node.val <= 5000

 

Follow up: A linked list can be reversed either iteratively or recursively. Could you implement both?


The idea is to keep 2 pointers for current and next node and then iterate through nodes and reconnect nodes. The best way to understand this solution is to take some small list, for example 1 -> 2 -> 3 -> 4 -> None and see what is going on on the each step.

    In the beginning curr = None, nxt = 1.
    On the first step we have: tmp = 2, nxt.next = None, curr = 1, nxt = 2. So what we have is the following: None <- 1 2 -> 3 -> 4 -> None. That is we cut one link between 1 and 2 and create new link 1 -> None.
    On the next step we have None <- 1 <- 2 3 -> 4 -> None and so on.

Try to draw it on the paper, step by step, it is the best way to feel it.

Complexity

Time complexity is O(n), space is O(1).