Basic Calculator
Given a string s representing a valid expression, implement a basic calculator to evaluate it, and return the result of the evaluation.

Note: You are not allowed to use any built-in function which evaluates strings as mathematical expressions, such as eval().

 

Example 1:

Input: s = "1 + 1"
Output: 2

Example 2:

Input: s = " 2-1 + 2 "
Output: 3

Example 3:

Input: s = "(1+(4+5+2)-3)+(6+8)"
Output: 23

 

Constraints:

    1 <= s.length <= 3 * 105
    s consists of digits, '+', '-', '(', ')', and ' '.
    s represents a valid expression.
    '+' is not used as a unary operation.
    '-' could be used as a unary operation but it has to be inside parentheses.
    There will be no two consecutive operators in the input.
    Every number and running calculation will fit in a signed 32-bit integer.

This is a very good stacks problem , we are given an expression in form of string with 
chars as [nums , '(' , ')' , '+' , '-' , ' ']

Note : we can totally ignore ' ' and  '+'.
    
Logic:
        what we are going to do is very simple , 
        first intilaize two vars , sum and sign
        sum = 0
        sign = +1
            
        traverse the string:
            if(ch is digit)
                num = get_full_num; // may be multidigit
                add it to sum , sum += (num * sign)
                reset sign to +1
            else if(ch is '(')
                save current state of sum and sign in stack
                reset sum and sign
            else if(ch is ')')
                sum = val_at_top + (sign_at_top * sum)
                pop;
            else if(ch is '-')
                toggle sign
                
       the use of sum and sign are to evaluate current expression inside a () braces or completely outside
       stack is used to store the state of sum and sign take example
       1 + 2 + 3 - (5 + 6) => 6 - (5 + 6)
                when we encounter '(' saving state means
                stack will store pair(6 , -) and first calculate (5 + 6)
                once (5 + 6) is calculated which is 11
                pop from stack => sum = (6 + (sum * -1)) = (6 + (11 * -1)) = -5 ans

or

This algorithm works for Basic Calculator (BC I) problem, where we can have only + - ( ) operations, for Basic Calculator II (BC II), where we can have only + - * / operations and also for Basic Calculator III (BC III), where we can have all + - * / ( ) operations.

Stack of monomials

The idea is to use both stack and recursion (which can be seen as 2 stack, because recursion use implicit stack). First, let us consider, that we do not have any brackets. Then let us keep the stack of monomial, consider the example s = 1*2 - 3\4*5 + 6. Then we want our stack to be equal to [1*2, -3\4*5, 6], let us do it step by step:

    Put 1 into stack, we have stack = [1].
    We can see that operation is equal to *, so we pop the last element from our stack and put new element: 1*2, now stack = [1*2].
    Now, operation is equal to -, so we put -3 to stack and we have stack = [1*2, -3] now
    Now, operation is equal to \, so we pop the last element from stack and put -3\4 instead, stack = [1*2, -3\4]
    Now, operation is equal to *, so we pop last element from stack and put -3\4*5 instead, stack = [1*2, -3\4*5].
    Finally, operation is equal to +, so we put 6 to stack: stack = [1*2, -3\4*5, 6]

Now, all we need to do is to return sum of all elements in stack.

How to deal with brackets

If we want to be able to process the brackets properly, all we need to do is to call our calculator recursively! When we see the open bracket (, we call calculator with the rest of our string, and when we see closed bracket ')', we give back the value of expression inside brackets and the place where we need to start when we go out of recursion.

Complexity

Even though we have stack and also have recursion, we process every element only once, so time complexity is O(n). However we pass slice of string as argument each time we meet bracket, so time complexity can go upto O(n^2) on example like (1+(1+(... +))) with O(n) open brackets. Space complexity is potentially O(n), because we need to keep stacks, but each element not more than once.

or

The problem of previous code is that we pass slice of string as parameter. In python it works quite fast, because function is implemented in C and it works very fast. If we want to have honest linear time, we need to pass index as parameter. (there is alternative way like I used in problem 1896 https://leetcode.com/problems/minimum-cost-to-change-the-final-value-of-expression/discuss/1267304/Python-Recursion-dfs-solution-explained, where we can precalculate pairs of open and closing brackets)

Complexity

Now time complexity it is O(n), space is still O(n).