Reachable Nodes In Subdivided Graph
You are given an undirected graph (the "original graph") with n nodes labeled from 0 to n - 1. You decide to subdivide each edge in the graph into a chain of nodes, with the number of new nodes varying between each edge.

The graph is given as a 2D array of edges where edges[i] = [ui, vi, cnti] indicates that there is an edge between nodes ui and vi in the original graph, and cnti is the total number of new nodes that you will subdivide the edge into. Note that cnti == 0 means you will not subdivide the edge.

To subdivide the edge [ui, vi], replace it with (cnti + 1) new edges and cnti new nodes. The new nodes are x1, x2, ..., xcnti, and the new edges are [ui, x1], [x1, x2], [x2, x3], ..., [xcnti+1, xcnti], [xcnti, vi].

In this new graph, you want to know how many nodes are reachable from the node 0, where a node is reachable if the distance is maxMoves or less.

Given the original graph and maxMoves, return the number of nodes that are reachable from node 0 in the new graph.

 

Example 1:
https://s3-lc-upload.s3.amazonaws.com/uploads/2018/08/01/origfinal.png
Input: edges = [[0,1,10],[0,2,1],[1,2,2]], maxMoves = 6, n = 3
Output: 13
Explanation: The edge subdivisions are shown in the image above.
The nodes that are reachable are highlighted in yellow.

Example 2:

Input: edges = [[0,1,4],[1,2,6],[0,2,8],[1,3,1]], maxMoves = 10, n = 4
Output: 23

Example 3:

Input: edges = [[1,2,4],[1,4,5],[1,3,1],[2,3,4],[3,4,5]], maxMoves = 17, n = 5
Output: 1
Explanation: Node 0 is disconnected from the rest of the graph, so only node 0 is reachable.

 

Constraints:

    0 <= edges.length <= min(n * (n - 1) / 2, 104)
    edges[i].length == 3
    0 <= ui < vi < n
    There are no multiple edges in the graph.
    0 <= cnti <= 104
    0 <= maxMoves <= 109
    1 <= n <= 3000

Here we need shortest distances, so Dijkstra algorithm is good idea. First of all, let us forgot about split points, and deal with graph as weighted graph, where weight of each node is number of split points on it plus one. On this graph we perform Dijkstra algorithm. Now at each point we have distance for it from origin. We can find total number of reachable points as:

    All original nodes with distance less or equal to M.
    For every edge we need to check what we have as distances for it ends:
    if both of distances more than M, than we can not reach any split points on this node. If both of them less or equal to M, we can reach some point starting with one end and some points, starting with another end, these two sets can intersect, so we need to remove then number of points in intersection. Also if for one end distance is more than M and for another is less or equal than M, we have only one set of points that can be reached.

Complexity

Time complexity is O(E log N), space complexity is O(N) as for classical Dijkstra algorithm.