initial commit
Tic-tac-toe is played by two players A and B on a 3 x 3 grid.

Here are the rules of Tic-Tac-Toe:

    Players take turns placing characters into empty squares (" ").
    The first player A always places "X" characters, while the second player B always places "O" characters.
    "X" and "O" characters are always placed into empty squares, never on filled ones.
    The game ends when there are 3 of the same (non-empty) character filling any row, column, or diagonal.
    The game also ends if all squares are non-empty.
    No more moves can be played if the game is over.

Given an array moves where each element is another array of size 2 corresponding to the row and column of the grid where they mark their respective character in the order in which A and B play.

Return the winner of the game if it exists (A or B), in case the game ends in a draw return "Draw", if there are still movements to play return "Pending".

You can assume that moves is valid (It follows the rules of Tic-Tac-Toe), the grid is initially empty and A will play first.

 

Example 1:

Input: moves = [[0,0],[2,0],[1,1],[2,1],[2,2]]
Output: "A"
Explanation: "A" wins, he always plays first.
"X  "    "X  "    "X  "    "X  "    "X  "
"   " -> "   " -> " X " -> " X " -> " X "
"   "    "O  "    "O  "    "OO "    "OOX"

Example 2:

Input: moves = [[0,0],[1,1],[0,1],[0,2],[1,0],[2,0]]
Output: "B"
Explanation: "B" wins.
"X  "    "X  "    "XX "    "XXO"    "XXO"    "XXO"
"   " -> " O " -> " O " -> " O " -> "XO " -> "XO " 
"   "    "   "    "   "    "   "    "   "    "O  "

Example 3:

Input: moves = [[0,0],[1,1],[2,0],[1,0],[1,2],[2,1],[0,1],[0,2],[2,2]]
Output: "Draw"
Explanation: The game ends in a draw since there are no moves to make.
"XXO"
"OOX"
"XOX"

Example 4:

Input: moves = [[0,0],[1,1]]
Output: "Pending"
Explanation: The game has not finished yet.
"X  "
" O "
"   "

 

Constraints:

    1 <= moves.length <= 9
    moves[i].length == 2
    0 <= moves[i][j] <= 2
    There are no repeated elements on moves.
    moves follow the rules of tic tac toe.

   Hide Hint #1  
It's straightforward to check if A or B won or not, check for each row/column/diag if all the three are the same.
   Hide Hint #2  
Then if no one wins, the game is a draw iff the board is full, i.e. moves.length = 9 otherwise is pending.

Not difficult problem, but in my opinion you need to code quite a lot for an easy problem. The idea is to keep number of X and O on each of vertical, horizontal and two diagonals. In fact we do not need to keep both of them, but do the following trick: we can add -1 to score if we meet X and +1 to score if we meet O. In this way game is finished if some score is equal to 3 or -3: if we have this numbers, it means, that we have full OOO of full XXX on the line.

So, we traverse moves one by one do the steps:

    Calculate player, it will be either 1 or -1 depending on parity of turn.
    Update rows[x], cols[y] and if needed diag and anti.
    Check if given move will finish the game: it can happen if one of the values rows[x], cols[y], diag, anti becomes equal to 3 or -3.
    In the end if we make 9 moves, we have Draw, in the opposite case it is Pending.

Complexity

If we have m moves, then we have O(m) time complexity. Space complexity is O(n), where here n = 3 is the size of our board.